Use application service for engine
