+++ resources/META-INF/plugin.xml
@@ -1424,6 +1424,7 @@
                         serviceImplementation="io.flutter.settings.FlutterSettings"
                         overrides="false"/>
 
+    <applicationService serviceImplementation="io.flutter.jxbrowser.EmbeddedBrowserEngine" overrides="false" />
     <console.folding implementation="io.flutter.console.FlutterConsoleFolding"/>
     <console.folding implementation="io.flutter.logging.FlutterConsoleLogFolding"/>
